reparat warninguri

+++ Album.cpp
@@ -32,7 +32,7 @@ Album::Album(const std::string &titlu,
              const int &an,
              const std::vector<Melodie> &melodii) : titlu{titlu}, an{an}, melodii{melodii} {
     lungime = 0;
-    for (int i = 0; i < getMelodii().size(); i++) {
+    for (size_t i = 0; i < getMelodii().size(); i++) {
         lungime += melodii[i].getLungime();
     }
 }
